+++ checkpoint/CHANGELOG.md
@@ -9,7 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Introduce Context and its usage as a contextmanager.
+- V1: Introduce Context and its usage as a contextmanager.
+- V1: Add `ArrayStorageOptions` to customize per-leaf saving behavior for
+arrays (e.g. `dtype`).
 
 ### Fixed